2007-07-31	Ralf Corsépius <ralf.corsepius@rtems.org>
	* score/inline/rtems/score/priority.inl: Use size_t instead of
	  uint32_t for array index.

+++ cpukit/score/inline/rtems/score/priority.inl
@@ -32,7 +32,7 @@
 
 RTEMS_INLINE_ROUTINE void _Priority_Handler_initialization( void )
 {
-  uint32_t   index;
+  size_t index;
 
   _Priority_Major_bit_map = 0;
   for ( index=0 ; index <16 ; index++ )
